-연료 계산이 잘못되던 부분 수정. -함선 정보가 고정값으로 되어있던 부분 유저인포에 따라 변하도록 수정. -워프 상태에서 아이템 스프라이트 추가. -스토리 -> 갤럭시 -> 시스템으로 씬 변경. -스테이션에서 함선수리, 연료채우기 동작시 골드가 없으면 취소되도록 수정.

+++ trunk/SpaceTrader2/src/com/raimsoft/spacetrader/data/UserInfo.java
@@ -319,7 +319,7 @@ public class UserInfo
 	
 	public int getCurrFuelPercent()
 	{
-		return (int) (nCurrFuel / nShipMaxFuel * 100.0f);
+		return (int) ((float)nCurrFuel / (float)nShipMaxFuel * 100.0f);
 	}
 	
 	/**
